removed getremotefile temporarily

+++ outgoingevents/downloadfile.go
@@ -1,10 +1,10 @@
 package outgoingevents
 
 type DownloadFile struct {
-	FileID      int32 `json:"file_id"`
-	Priority    int32 `json:"priority"`
-	Offset      int32 `json:"offset"`
-	Limit       int32 `json:"limit"`
+	FileID      int64 `json:"file_id"`
+	Priority    int64 `json:"priority"`
+	Offset      int64 `json:"offset"`
+	Limit       int64 `json:"limit"`
 	Synchronous bool  `json:"synchronous"`
 }
